Add keyboard shortcut to end mentoring for mentors
