+++ package.json
@@ -40,6 +40,7 @@
     "humps": "^2.0.1",
     "localforage": "^1.9.0",
     "lodash": "^4.17.20",
+    "mousetrap": "^1.6.5",
     "nim-codemirror-mode": "^0.2.0",
     "pluralize": "^8.0.0",
     "prop-types": "^15.7.2",
@@ -79,6 +80,7 @@
     "@types/fontfaceobserver": "^0.0.6",
     "@types/jquery": "^3.5.5",
     "@types/lodash": "^4.14.165",
+    "@types/mousetrap": "^1.6.8",
     "@types/pluralize": "^0.0.29",
     "@types/qs": "^6.9.5",
     "@types/react": "^16.9.51",
